Enriched order object with data

+++ routes/OrderRoute.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const {getResponse} = require('../helpers');
+const {getResponse, prepareOrders} = require('../helpers');
 
 // order model
 const Order = require("../models/Order");
@@ -9,7 +9,8 @@ const Order = require("../models/Order");
 // returns all order items
 router.get("/", async (req, res) => {
     try{
-        const orders = await Order.find();
+        let orders = await Order.find();
+        orders = await prepareOrders(orders);
         return res.status(200).json(getResponse(orders, 'Success'));
     } catch(err){
         console.log(err);
@@ -21,9 +22,10 @@ router.get("/", async (req, res) => {
 // returns one order found by id
 router.get("/:orderId", async (req, res) => {
     try{
-        const order = await Order.findById(req.params.orderId);
+        let order = await Order.findById(req.params.orderId);
         if(order){
-            return res.status(200).json(getResponse(order, 'Success'));
+            order = await prepareOrders([order]);
+            return res.status(200).json(getResponse(order[0], 'Success'));
         } else {
             return res.status(404).json(getResponse(null, 'Not Found'));
         }
@@ -36,10 +38,11 @@ router.get("/:orderId", async (req, res) => {
 //POST
 //creates new order
 router.post("/", async (req, res) => {
-    /*if(!req.logged){
+    if(!req.logged){
         return res.status(403).json(getResponse(null, 'Unauthorized'));
-    }*/
+    }
 
+    // enrich
     const newOrder = new Order({
         pollId: req.body.pollId,
         restaurantId: req.body.restaurantId,
@@ -63,27 +66,40 @@ router.post("/", async (req, res) => {
 //edit
 //change order by id
 router.put("/:orderId", async (req, res) => {
-  try {
-    const foundOrder = await Order.findById(req.params.orderId);
-    foundOrder.status = req.body.status;
-    foundOrder.orderItemsList = req.body.orderItemsList
-    res.json(foundOrder);
-  } catch (err) {
-    res.status(500).json({ message: err });
-  }
+    if(!req.logged){
+        return res.status(403).json(getResponse(null, 'Unauthorized'));
+    }
+
+    try {
+        const order = await Order.findOneAndUpdate({_id: req.params.orderId}, { ...req.body }, {useFindAndModify: false});
+        if(order){
+            return res.status(200).json(getResponse({ ...order._doc, ...req.body }, 'Success'));
+        } else {
+            return res.status(404).json(getResponse(null, 'Error while editing poll'));
+        }
+    } catch(err){
+        console.log(err);
+        return res.status(500).json(getResponse(null, err));
+    }
 });
 
 // delete
 router.delete("/:orderId", async (req, res) => {
-  try {
-    const deletedOrder = await Order.findByIdAndDelete(req.params.orderId);
-    res.json({
-      order: { ...deletedOrder._doc },
-      message: "Success",
-    });
-  } catch (err) {
-    console.log(err);
-  }
+    if(!req.logged){
+        return res.status(403).json(getResponse(null, 'Unauthorized'));
+    }
+
+    try {
+        const deletedOrder = await Order.findByIdAndDelete(req.params.orderId);
+        // izbrisi sve orderiteme iz deletedOrdera
+        if(deletedOrder){
+            return res.status(200).json(getResponse(null, 'Success'));
+        } else {
+            return res.status(404).json(getResponse(null, 'Not Found'));
+        }
+    } catch (err) {
+        console.log(err);
+    }
 });
 
 module.exports = router;
